Add back reverted changes for short segment acceleration  (with fixes).
The small circles moving very fast was due to the move segment speed
getting set early to the "desired" target speed instead of the real
target speed after the acceleration calcs.

The straight segement hang was being caused by the "fudging" code that
ignored the acceleration segment when the move was very short. This was
causing a "move" speed of 0 when the "last" speed seen by the planner
was 0.

Signed-off-by: H Hartley Sweeten <hsweeten@visionengravers.com>